Refactor filter expression parser
